move Carousel custom render functions to src/utils/Carousel.js
